requirements.txt
httpx>=0.27
beautifulsoup4>=4.12
lxml>=5.1
jinja2>=3.1
playwright>=1.40
apscheduler>=3.10
pyyaml>=6.0
python-dotenv>=1.0
loguru>=0.7
